feat: restructure code to atomic design
