+++ src/common/templates/App/index.tsx
@@ -1,6 +1,5 @@
 import { RouterProvider } from 'react-router-dom';
 import { ServiceProvider, services } from 'src/services';
-// import './App.css';
 import { router } from 'src/router';
 
 export function App() {
