--- src/demo/lazy-load-image/index.tsx
@@ -1,2 +1,0 @@
-export { LazyLoadImageDemo1 } from './demo-1';
-export { LazyLoadImageDemo2 } from './demo-2';
