+++ src/demo/pages/ImageLazyLoad/demo-2.tsx
@@ -10,7 +10,7 @@ function ImageGallery({ images }: { images: string[] }) {
   );
 }
 
-export function LazyLoadImageDemo2() {
+export function ImageLazyLoad2() {
   const images = [
     'https://picsum.photos/id/1/300/300',
     'https://picsum.photos/id/2/300/300',
