+++ src/demo/pages/ImageLazyLoad/index.tsx
@@ -1,0 +1,2 @@
+export { ImageLazyLoad1 as LazyLoadImageDemo1 } from './demo-1';
+export { ImageLazyLoad2 as LazyLoadImageDemo2 } from './demo-2';
